Dev

Approved-by: Mathieu Boisnard <boisna_m13@esgi.fr>

+++ tci-projects/src/main/java/com/travelci/projects/services/ProjectsServiceImpl.java
@@ -1,8 +1,11 @@
 package com.travelci.projects.services;
 
-import com.travelci.projects.adapter.ProjectAdapter;
+import com.travelci.projects.entities.PayLoad;
+import com.travelci.projects.entities.ProjectAdapter;
 import com.travelci.projects.entities.ProjectDto;
+import com.travelci.projects.exceptions.NotFoundProjectException;
 import com.travelci.projects.repository.ProjectRepository;
+import org.eclipse.jgit.api.Git;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,13 +18,15 @@ import java.util.stream.Collectors;
 public class ProjectsServiceImpl implements ProjectsService {
 
     private final ProjectRepository projectRepository;
-
     private final ProjectAdapter projectAdapter;
+    private final GitService gitService;
 
     public ProjectsServiceImpl(final ProjectRepository projectRepository,
-                               final ProjectAdapter projectAdapter) {
+                               final ProjectAdapter projectAdapter,
+                               final GitService gitService) {
         this.projectRepository = projectRepository;
         this.projectAdapter = projectAdapter;
+        this.gitService = gitService;
     }
 
     @Override
@@ -31,6 +36,14 @@ public class ProjectsServiceImpl implements ProjectsService {
             .stream()
             .map(projectAdapter::toProjectDto)
             .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public ProjectDto getProjectById(final Long projectId) {
+        return projectAdapter.toProjectDto(
+            projectRepository.findById(projectId).orElseThrow(NotFoundProjectException::new)
+        );
     }
 
     @Override
@@ -46,18 +59,42 @@ public class ProjectsServiceImpl implements ProjectsService {
 
     @Override
     public ProjectDto update(final ProjectDto projectDto) {
+
+        if (projectRepository.findOne(projectDto.getId()) == null)
+            throw new NotFoundProjectException();
+
         projectDto.setUpdated(new Timestamp(System.currentTimeMillis()));
-        return projectAdapter.toProjectDto(projectRepository.save(projectAdapter.toProjectEntity(projectDto)));
+
+        return projectAdapter.toProjectDto(
+            projectRepository.save(projectAdapter.toProjectEntity(projectDto))
+        );
     }
 
     @Override
     public void delete(final ProjectDto projectDto) {
+
+        if (projectRepository.findOne(projectDto.getId()) == null)
+            throw new NotFoundProjectException();
+
         projectRepository.delete(projectAdapter.toProjectEntity(projectDto));
     }
 
     @Override
     @Transactional(readOnly = true)
-    public ProjectDto getProjectDetails(final Long projectId) {
-        return projectAdapter.toProjectDto(projectRepository.getOne(projectId));
+    public void startProjectEngine(final PayLoad webHookPayLoad) {
+
+        // Check incoming webhook with project list
+        final ProjectDto searchProject = projectAdapter.toProjectDto(
+            projectRepository.findFromPayLoad(
+                webHookPayLoad.getRepositoryUrl(),
+                webHookPayLoad.getBranchName()
+            ).orElseThrow(NotFoundProjectException::new)
+        );
+
+        // Get Source Code
+        final Git repository = gitService.pullProjectRepository(searchProject, webHookPayLoad);
+        repository.close();
+
+        // Send Request to tci-commands
     }
 }
